cleanup: refactored code with PR review
