+++ scrapbook-gateway-service/app.py
@@ -5,7 +5,6 @@ import sys
 import requests
 
 from routes.auth import authenticate_user_api
-from routes.session import logout_user_api
 from routes.album import album_api
 from routes.image import image_api
 
@@ -15,13 +14,13 @@ app = Flask(__name__)
 # Cross Origin Resource Sharing (CORS) is enabled for the app
 CORS(app)
 
-#retries user info from the user service, and adds the use to a new session.
+#contains the auth api services 
 app.register_blueprint(authenticate_user_api)
 
-#manuoppilating album information from datavbase    
+#manupilating album information using the image service    
 app.register_blueprint(album_api)
 
-#download/uploads an image
+#download/uploads an image using Gdrive service
 app.register_blueprint(image_api)
