+++ scrapbook-gateway-service/routes/auth.py
@@ -2,7 +2,6 @@ import requests
 from flask import Blueprint, request, jsonify
 import sys
 from service_utils import auth_service as auth
-
 from config import USER_SERVICE_URL__DEV, SESSION_SERVICE_URL__DEV 
 
 authenticate_user_api = Blueprint('authenticate_user_api', __name__)
@@ -26,6 +25,25 @@ def login():
         session_response = requests.post(f'{SESSION_SERVICE_URL__DEV}/set', data = {"userID": user['email'], "token": user['token']})
         session_response.raise_for_status()
         return response.json(), response.status_code
+
+    except requests.exceptions.HTTPError as err:
+        return err.response.text, err.response.status_code
+
+
+@authenticate_user_api.route('/logout', methods=["GET"])
+@auth.checkUserSession()
+def logout():
+    """
+    Removes user data from the session
+
+    @params - Delete request sent to the session service with session id as a parameter
+    @return - OK if removed from the session db and http status code
+    """
+    try:
+        userID = request.values['id']
+        response = requests.delete(f'{SESSION_SERVICE_URL__DEV}/remove/{userID}')
+        response.raise_for_status()
+        return response.content, response.status_code
 
     except requests.exceptions.HTTPError as err:
         return err.response.text, err.response.status_code
